open-library-main/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
aiofiles
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pytest
httpx
